requirements.txt
streamlit
google-generativeai
python-dotenv
requests
pillow
pandas
streamlit-option-menu
streamlit-extras